Trying to create an calculate activity that calculate the value of the activity multiplied with category multiplier

+++ Components/Classes/EducationCategory.cs
@@ -4,7 +4,7 @@
     {
         public string CategoryName { get; set; } = string.Empty;
         public string Description { get; set; } = string.Empty;
-        public int HoursMultiplier { get; set; } = 1;
+        public float HoursMultiplier { get; set; } = 1;
 
     }  
 }
